Refine the public APIs for IPFS storage (#175)

+++ ElastosHiveSDK/Vendors/IPFS/IPFSRpc.swift
@@ -76,7 +76,7 @@ class IPFSRpc: ConnectHelper {
         connectState = true
     }
     
-    private func checkConnect(_ ip: String?, _ port: UInt16) -> Bool {
+    private func checkConnect(_ ip: String?, _ port: Int) -> Bool {
         guard ip != nil else {
             return false
         }
